Revert leave session functionality

+++ frontend/src/pages/CollaborationPage.jsx
@@ -64,7 +64,9 @@ const CollaborationPage = () => {
 
   const handleEndSession = () => {
     setShowModal(true);
+    /*  Feature Disabled
     socket.emit('confirmEndSession', sessionId, secondsElapsed);
+    */
   };
 
   const findBestAttempt = (attempts) => {
@@ -99,6 +101,7 @@ const CollaborationPage = () => {
     return bestAttempt
   };
 
+  /*  Unused - replaced with better method
   const formatTimeTaken = () => {
     const hours = Math.floor(secondsElapsed / 3600);
     const minutes = Math.floor((secondsElapsed % 3600) / 60);
@@ -109,17 +112,20 @@ const CollaborationPage = () => {
     } else {
       return `${minutes}:${seconds}`;
     }
-  }
+  }*/
 
   const handleConfirmEndSession = () => {
+    console.log("Attempt to leave session.")
     socket.emit('endSession', sessionId, secondsElapsed);
     setShowModal(false);
+    endSessionAndNavigate(secondsElapsed);
   };
 
   const handleCancelEndSession = () => {
     setShowModal(false);
   };
 
+  /*  Disabled - feature disabled
   const handleUser2Response = (response) => {
     setIsOtherUserPrompted(false);
     if (response) {
@@ -129,20 +135,25 @@ const CollaborationPage = () => {
       // Emit cancellation if other user disagrees
       socket.emit('responseEndSession', sessionId, 'no');
     }
-  };
+  };*/
 
+  /*  Disabled - feature disabled
   useEffect(() => {
     socket.on('confirmEndSession', (elapsedTime) => {
+      
       if (!isOtherUserPrompted) {
         setIsOtherUserPrompted(true);
         setSecondsElapsed(elapsedTime);
       }
+      toast.info('Your partner has left the session.', { autoClose: false }); // User has to close notification manually
     });
 
+    
     socket.on('endSessionBoth', (elapsedTime) => {
       endSessionAndNavigate(elapsedTime);
     });
 
+    
     socket.on('userContinues', () => {
       toast.info('Your partner has chosen to continue the session.', { duration: 10000 }); // display for 10s
     });
@@ -152,7 +163,7 @@ const CollaborationPage = () => {
       socket.off('endSessionBoth');
       socket.off('userContinues');
     };
-  }, [sessionId, isOtherUserPrompted]);
+  }, [sessionId, isOtherUserPrompted]);*/
 
   const endSessionAndNavigate = (elapsedTime) => {
     //     socket.emit('endSession', sessionId);
@@ -184,12 +195,13 @@ const CollaborationPage = () => {
         });
       })
       .then(() => {
-        fetch(`http://localhost:8084/sessions/${sessionId}`, {
-          method: 'DELETE',
-          headers: { 'Content-Type': 'application/json' },
-          body: JSON.stringify({ sessionid: sessionId }),
-        });
-
+        if (localStorage.getItem('partnerLeft')) {
+          fetch(`http://localhost:8084/sessions/${sessionId}`, {
+            method: 'DELETE',
+            headers: { 'Content-Type': 'application/json' },
+            body: JSON.stringify({ sessionid: sessionId }),
+          });
+        }
         navigate('/summary', { state: { matchData, secondsElapsed: elapsedTime } });
       })
       .catch((error) => console.error("Error ending session:", error));
@@ -219,14 +231,14 @@ const CollaborationPage = () => {
         onConfirm={handleConfirmEndSession}
         onCancel={handleCancelEndSession}
       />
-
-      {/* Other user's confirmation modal */}
+      
+      {/* Other user's confirmation modal 
       <ConfirmationModal2
         show={isOtherUserPrompted}
         message="Your partner wants to end the session. Do you agree?"
         onConfirm={() => handleUser2Response(true)}
         onCancel={() => handleUser2Response(false)}
-      />
+      />*/}
 
       {/* Main Content Section */}
       <div style={{ display: 'flex', flex: 1, flexDirection: 'row', gap: '20px', marginTop: '70px', width: '100%', height: '700px' }}>
